Support decimal hours for archive threshold
Allow users to set archive times with decimal precision for finer control:
- Change input from step="1" to step="0.1" (6 minute increments)
- Change minimum from min="1" to min="0.1" (6 minute minimum)
- Change parseInt() to parseFloat() to preserve decimal values
- Update help text with examples (0.5 = 30 min, 0.25 = 15 min)

Examples now supported:
- 0.5 hours = 30 minutes
- 0.25 hours = 15 minutes
- 1.5 hours = 1 hour 30 minutes

+++ options.js
@@ -79,7 +79,7 @@ async function loadContainers() {
 async function saveSettings() {
   const settings = {
     archiveEnabled: document.getElementById('archiveEnabled').checked,
-    archiveAfterHours: parseInt(document.getElementById('archiveAfterHours').value),
+    archiveAfterHours: parseFloat(document.getElementById('archiveAfterHours').value),
     excludePinnedTabs: document.getElementById('excludePinnedTabs').checked,
     excludedDomains: getCurrentExcludedDomains(),
     workspaceRules: getCurrentWorkspaceRules(),
